rename 'dq_utils' directory to 'utils'
